"11am. I went to bed far too late yesterday.
11:05am. Let me read the OPM dump and then I'll do some work on the docs.

11:20am. Let me work on the docs for an hour. I need to get into this for a bit. It is time to start wrapping things up.

///

The previous segment is not all there is to serialization. I could have picked other subjects to demonstrate the language's prowess at abstraction, but serialization is a common task that other languages rarely do without flaw. It will only get more important as time goes on. When I picked F# as main target in the old version of Spiral it was mistake because I had no automatic way to manage GPU memory. Yet, the future hardware will not have unified memory and it won't be necessary to deal with unmanaged pointers like in the past. Instead it will all be about message passing between isolated cores, and being able to succinctly and efficiently serialize data in various formats and having it cross various kinds of boundaries is Spiral's greatest strength.

Back in 2016, I ran into a fairly extreme serialization challenge when trying to pass data to a NN agent. At that time, I had no idea how to do pickler combinators so my situation was that much worse - I fervently wished for some way to map a regular type to a sequence of one-hot vectors.

An one-hot vector is an array whose elements are all 0, apart from a single element which is 1. I needed an efficient way to do this task - I tried doing it by hand in C style, but the problem with doing that quickly becomes aparent. It is extemely unsafe and very hard to do. It is a killer combination that slowly set me on the path of being a language designer.

Imagine you are doing a NN-based poker agent, and the you are trying to pass an input to its model. The is a simple HU poker variant with two cards. Here is how the game state could be modeled.

```
type rank = i32
type suit = i32
type card = rank * suit
type player = {
    stack : i32
    hand : card * card
    }
type pot = i32
type game = player * player * pot
```

We'll assume that the opponent's previous action could be inferred from the size of the pot. But the player should receive his previous action during this betting round. So the actual player input would be something like this.

```
union action =
    | Raise: i32
    | Call
    | NoAction
type player_view = {
    stack_self : i32
    stack_opp : i32
    pot : i32
    hand : card * card
    prev_action : action
    }
```

This `player_view` is what needs to be serialized. Compared to the previous scheme, things are more difficult this time. There are two main challenges.

1) Serializing arbirary ints is far too wasteful. An int converted to a one hot-vector would take 2Gb. Instead one has to account for maximum and minimum stack sizes which might be between 0 to a few 100 chips. Suits are only 4, and the number of different cards is 13. In other words, the problem is that unlike in the previous section, the type no longer corresponds to the representation being used. It would be great if Spiral had dependent types so that bounded ints could be used, but that is not the case here. There is no problem with using ints for stack sizes and such inside the game itself, but they are huge problem to serialize in the one-hot format.

2) Dealing with sizes. As in the previous section, the union types are tricky here because one case is prefixed by the rest.

---

Because of the added difficulties, unlike when it comes to vanilla serialization, the introspective approach is no longer so obviously superior to the introspective one. If all the bounds are known at compile time, an introspective approach would involve using nominals with symbolic bounds which would require hackery like converting symbols to ints to serve as bounds. I previously had the idea of specifying the size using a prototype function, but something like `prototype size a : i32` is not allowed in Spiral because `a` is an unused forall variable.

In the old version of Spiral, I did not use pickler combinators, but I still had to specify a schema even there, so the advantages of the introspective approach is minimal over the combinator one.

///

Let me back this up here. I'll redo this from the start.

11:40am. I see that I got a reply from somebody trying out Spiral on Linux. Forget the docs for the time being.

///

> However, when I try to use it, I see a notification: "Spiral: The server has aborted with an error."

This happens when the VS Code plugin runs the server process, but the process aborts for some reason.

///

12:35pm. Let me have breakfast here. I opened an issue on the Spiral repo and put down my thoughs. For the rest of the day, I'll focus on dealing with the last segment. After that I'll clear up some space on my hard drive and get Linux to install."