"10:30am. Let me chill for a while.
11am. Let me start here.

There is no avoiding this. Let me just start work on the comment. It might take me a day or a few, but it will be done. After that I am will be done with challenging work on the compiler. I will get back to the docs after that.

```fs
comments : LineComment option []
```

It is finally time to make use of this.

For now, let me forget the later phases. Let me just concern myself with the parser.

11:10am.

```fs
    | TopInl of Comments * VSCRange * (VSCRange * VarString) * RawExpr * is_top_down: bool
    | TopRecInl of Comments * VSCRange * (VSCRange * VarString) * RawExpr * is_top_down: bool
```

Ok, I've sprinkled this in all the right places.

The parser is done.

Let me continue going forward.

```fs
type TopEnv = {
    nominals_next_tag : int
    nominals_aux : Map<GlobalId, {|name : string; kind : TT|}>
    nominals : Map<GlobalId, {|vars : Var list; body : T|}>
    prototypes_next_tag : int
    prototypes_instances : Map<GlobalId * GlobalId, Constraint Set list>
    prototypes : Map<GlobalId, {|name : string; signature: T|}>
    ty : Map<string,T>
    term : Map<string,T>
    constraints : Map<string,ConstraintOrModule>
    }
```

`term` here will have to change so it includes the comments as well.

```fs
| TyModule of Map<string, Comments * T>
```

`TyModule` as well.

```fs
let in_module m a =
    {a with
        ty = Map.add m (TyModule a.ty) Map.empty
        term = Map.add m (TyModule a.term) Map.empty
        constraints = Map.add m (M a.constraints) Map.empty
        }
```

Argh, no, this is not it.

```fs
    | TyModule of Map<string, Comments * T>
    | TyFun of T * T
```

Instead of the module here, should I attack a comment to the `TyFun` itself?

```fs
    | TyModule of Map<string, T>
    | TyComment of Comments * T
    | TyFun of T * T
```

How about I go with this?

```fs
let rec tt (env : TopEnv) = function
    | TyComment(_,x) | TyMetavar(_,{contents=Some x}) -> tt env x
```

I'll have to adjust everything, but it will be worth it.

```fs
let rec visit_t = function
    | TyMetavar(_,{contents=Some x} & link) -> shorten x link visit_t
    | a -> a
```

Should I have visit ignore the comments? Hmmm, let me just roll with it like this for now.

Actually, for now let me not even focus on displaying the comments. Let me just add all the machine and make sure that things still work.

```fs
            match visit_t a'', visit_t b'' with
            | TyComment(_,a), b | a, TyComment(_,b) -> loop (a,b)
```

Dealing with these in unification is easy enough.

11:40am. Ok, the compiler builds. Let me try running it. But though I've filled in the inexhuastive cases, `TyComments` are not being used anywhere, so I expect this will work.

The way I've set things up now would allow local statement to propagate their comment data...

But I've made comment stripping really fast in union, so nevermind that.

Ah, yes. I should not ignore comment in show_t. That is what hover is using to print them.

I'll get to that later. Let me try running it. I just need to make sure that nothing is broken before I proceed.

```
Server bound to: tcp://*:13805 & tcp://*:13806
Unhandled exception: System.IndexOutOfRangeException: Index was outside the bounds of the array.
   at Spiral.BlockParsing.loop@1180-68(Env s, Int32 line, FSharpList`1 d) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language 2\BlockParsing.fs:line 1181
   at Spiral.BlockParsing.comments[a](Env s) in C:\Users\Marko\Source\Repos\The
Spiral Language\The Spiral Language 2\BlockParsing.fs:line 1186
   at Spiral.BlockParsing.top_inl_or_let@1208-17.Invoke(Env d) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language 2\BlockParsing.fs:line 1208
   at Spiral.BlockParsing.top_inl_or_let(Env d) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language 2\BlockParsing.fs:line 1208
   at Spiral.BlockParsing.top_statement@1257.Invoke(Env d) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language 2\BlockParsing.fs:line 1257
   at Spiral.BlockParsing.op_Addition@1256-5.Invoke(FSharpFunc`2 a, FSharpFunc`2 b, Env d) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language 2\BlockParsing.fs:line 1256
```

I did not expect this, but it is a good thing that I decided to be cautious and did the check.

```fs
if line < 0 then
```

What the hell is this.

```fs
let comments (s : Env) =
    let line_near_to = line s
    let rec loop line d =
        if 0 <= line then
            match s.comments.[line] with
            | Some(r,text) -> loop (line-1) (text :: d)
            | _ -> d
        else d
    loop (line_near_to-1) []
    |> String.concat "\n"
    |> fun x -> Ok(x.TrimEnd())
```

Let me go with this.

11:50am. Let me take a break here. I feel like I will be able to make rapid progress on this now that I've broken the ice.

I'll do the breakfast and chores, and then get back to this and deal with this today. Then I will be free."