"4:10pm. Done with the break. I've made it this far. Now let me start with proofreading the docs. This will not be hard work.
4:15pm. Focus me, let me just start.

https://chrome.google.com/webstore/detail/grammar-and-spell-checker/oldceeleldhonbafppcapldpdifcinji?utm_source=lt-homepage&utm_medium=referral

I guess I'll try installing this.

4:20pm. No I do not need a Chrome extension. I'll just go page by page. I'd rather prefer to be able to paste eveything in, but there is a 10k character limit for the free version of LanguageTools.

4:45pm. Things are going quickly. I am over third of the way through the docs. I'll have to proofread it top to bottom rather than just copy into the language tool as well.

5:25pm. The first part of proofreading is done. I've put the entire document through the language tool and fixed whatever errors I've come across. Now I need to actually read it myself once.

I am not really good at this kind of work though. Still let me do it.

Today I do this, and tomorrow the long war will start. For the first time in a long while, I won't be doing programming. I will start things off by making a list of companies to apply to.

6:20pm. Found a minor error in an error message for closure conversion so I fixed it. I'll publish the compiler again.

6:30pm. Done with proofreading. I caught a few errors, but my focus was low throughout. As I said, I am not good at proofreading my own work, I really rushed through this. I need a different pair of eyes to go through the docs in order to ensure their integrity.

The minor effort I put in now should suffice though. I got rid of all the really obvious mistakes.

This is a great spot to stop for the day. I am done with the docs with two weeks left in the month. Right now it is time to start gearing up for negotiations.

This will be quite a change of pace. I am looking forward to it."