"11am. http://www.spiral.net/doc/pdf/spiral-tutorial2019.pdf
I got reply that there is some name collision between my Spiral and another. I knew there was a DSL by the same name, but I thought it was a Scala LMS library.

I added a note in the readme. But this is a bit interesting, let me take a look at the tutorial for this.

After that I'll get started on the videos.

11:15am. Let me finish the thread I am lurking and then I'll get started with watching videos.

11:40am. That ended up being longer than I thought. Ah, well I always waste an hour after I get up. Let me start the vids. I'll start with Cornami.

https://www.youtube.com/watch?v=_bNzxW90AEI
Cornami: Next Generation Machine Learning Algorithms for Privacy Preserving Neural Networks

https://cornami.com/

Why are they talking about encription? I guess I'll know after I watch the video.

11:50am. https://youtu.be/_bNzxW90AEI?t=518

This is a pretty good technical talk. Ok, by this point I can see these guys are serious. I'll apply at Cornami amongst the others.

https://youtu.be/_bNzxW90AEI?t=722

It is interesting that they are worried about quantum computers.

https://youtu.be/_bNzxW90AEI?t=796
Fully Homomorphic Encryption: A holy grail of cryptography

https://youtu.be/_bNzxW90AEI?t=1155

This is actually pretty interesting.

12:25pm. Reconfigurable cores would definitely be up my alley. If I had one of these chips, making the RL agents that I want would not be a problem. Though I'd expect them to be hard to program. Still, it would be worth mastering this.

There are just so many of these companies. On the surface, my plan is to do a wide sample, but it might turn out to be too difficult to do in a timely manner. This might turn my choices into a coin flip.

In that case, I should just pick something like Cornami - adaptable hardware would adapt to any possible situation. And unlike say spiking nets, I won't be held back by the right algorithms not existing.

Let me move on to the next thing on the list.

The next one is Esperanto. Enflame is the only company which does not have a presence so I'll remove it from consideration.

https://www.youtube.com/watch?v=g9NawVqm3qM
RISC-V Tokyo 2020 : Shankar Jayaratnam, Raymond Tang | Esperanto Technologies (USA)

https://www.esperanto.ai/

Let me watch this vid. I'll do it while I have breakfast.

1:10pm. The Esperanto vids are low quality.

I'll say that Cornami is high ranked and Esperanto is mid at best.

In that one video, the japanese CTO started off by displaying his poor grasp of English. Things like this reveal a level of disfunction in the company. The video ended up being about hardware verification too instead of the chip itself. The other video is similar to that.

Forget this. Let me move on to the next thing.

The next is Fathom Computing. [Website](https://www.fathomcomputing.com/).

Damn it, not only can't I not find videos for this, I am having difficulty finding the team.

https://www.wired.com/story/this-computer-uses-lightnot-electricityto-train-ai-algorithms/

Here is an article on it. let me read this.

1:40pm. Fathom is too much in the research phase most likely. Forget it. Let me move on.

The next is Groq.

https://www.youtube.com/watch?v=uJbTEXFAsLs
Groq's Tensor Streaming Processor

https://youtu.be/uJbTEXFAsLs?t=73

Uncompromised peak performance for batch size 1 workloads.

I definitely need this.

2:15pm. Groq is high tier. It is not as speculative as Cornami. Adapting hardware might be hard to deal with so I will be taking a risk there that I would not with Groq.

Let me have a break here.

After that I'll watch more of Groq's videos."