"10:45am. I am up. Let me chill a little and I will start. Not programming, but opening that sponsor page on the Spiral repo. That has been a long time coming.
I noticed last night I got a reply by scoder on the TCO issue so I need to reply to him.

Let me do my morning reading first.

11:25am.

///

> I doubt that you would see any difference whatsoever in the runtime if that C level optimisation was applied to your code.

Do you mean to say that on the toy example the performance will be dominated by the memory allocation? If that is so, I understand that. If TCO worked though, the program would not stack overflow regardless of the number of iterations.

> You are asking for tail call optimisation after (probably) just creating a new Python object, allocating memory for it and initialising it?

I am pointing out that any new objects passed in a tail call effectively block TCO because they need to be decref'd after the method returns. For TCO to work, the method call needs to be the last and not followed by a decref.

For this to be the case, Cython needs an alternative construct to the usual `cdef` methods. This is what I am asking for.

If I were doing a C backend I could fulfil all of the requirements on the Spiral wishlist for it. I could do the same with Cython if I were to compile objects to pointers and do ref counting manually, but I do not know how to deal with exceptions. Exceptions would leave pointers hanging. So I have no choice, but to make a request.

///

Let me do a reply...this is good.

11:50am. Had to take a short break. I really spent my time slacking this morning. Before I stop for breakfast, I can get a bit work done. Let me see if that display card works. Let me crack open the envelope.

11:55am. Yeah, the card is there. This one does not seem to have an expiry date.

12:10pm. Logged in. Between 2015 and now, it seems I wasted 72$ dollars just for having the account open. Those thieves took 1.2$ per month from me. My father should have drained it completely.

Well, no matter. Let me see about opening that sponsor page.

https://docs.github.com/en/github/supporting-the-open-source-community-with-github-sponsors/receiving-sponsorships-through-github-sponsors

12:20pm. How disappointing. I joined the waiting list as Github Sponsors is not supported in Croatia.

Back in my trading days I had an account with Interactive Brokers, so that might have sufficed, but that has been closed for 7 years now. Should I be aiming to open an account outside the country just to get on this program? I am not sure.

I do not have much interest in this.

12:25pm. Forget this. It would have been different if I had an waiting audience for Spiral, but I'll just focus on my current plan instead.

Today I will change i32 to u32 arrays. I'll rename the modules to from `arrayu64` to `a64` and `arrayu32` to `a32`. It is really hard to keep writing out `arrayu64` every time. `array` or `list` are fine, but `arrayu64` are like 3 different words for me.

`a64` would be 2, but at least it will be concise.

12:30pm. https://cython.readthedocs.io/en/latest/src/userguide/language_basics.html#error-return-values

Oh, damn it. So this is how exceptions work in Cython. I actually hadn't even read this part of the docs. It seems that for every function that returns a primitive, I should be using except *.

Would this crush TCO? What would the performance impact of this be for that TCO loop?

```
cdef int loop_tail(int s, int i, int nearTo) except *:
    if i < nearTo: return loop_tail(s+i,i+1,nearTo)
    else: return s
cpdef int sequence_tailrec(int fron, int nearTo) except *:
    return loop_tail(0,fron,nearTo)
```

...Yes, it ends up blocking TCO just as one would expect.

1pm. Hmmm...maybe I will add a compiler option to generate `except *` on every join point. No way around this one. I really am learning a decent amount aren't I?

1:05pm. Let me have breakfast here. I am overdoing it."