"10:10am. I am up. Let me read Ayakashi and AssPride and then I will start.
I keep plotting how to get that point of contact and the only thing my mind is producing are bigger and bigger lies. The right approach for this is to fake identity and gather a lot of negotiation experience that way. I mean, imagine you are an AI. You could just make up fake resumes and apply through them.

The fact that I am not looking for a job, but a sponsor for Spiral actually makes things a lot harder for me. Spiral is tied to my identity and I can't really use the full breadth of negotiation strategies. Plus this is wasting a lot of my time and mental energy.

I actually resent a company like Groq a little. In the job posting there was a lot of what the company itself wants, but not what they would provide to the candidate. This goes even for senior posts.

Do I really have to approach this as a conman? If I get ghosted by the end of the week, I'll burn my morals and approach this solely from the perspective of maximizing my odds. I'll apply to senior positions from here on out with a bunch of made up experience.

10:15am. But I am really tempted to just wash my hands of this.

Forget the outside world and change my goals.

Yesterday I said that making a ML library was a mistake. That might have been the truth.

I am going to change my desired specialty towards where I am going. In 2020 I've started changing my approach from ML to concurrency. It should have been my main path from the start.

10:25am. Yes, making a ML library was a mistake. But I don't think that creating Spiral was. It gave me unparalleled freedom.

Like, once I had v2 it allowed me to simply make a Cython backend. That shows just how strong Spiral's capabilities are. Now I can program Python in the true functional style.

10:30am. I should be getting better at concurrency. More and more. Everything else is trivial. Even ML itself is trivial. It is a bunch of compiler engineering problems surrounded by NP Hard, impossible tier problems.

How difficult it is entirely depends on my approach.

...

I can get better endlessly if I cultivate my concurrency skills. It is really the natural outgrowth of functional programming. The fact that I thought that ML skills were the most central in the path of transcendence was merely just my misconception.

I might have made a tactical error, but the strategic choice of striving to overcome my limitations through making a language was the right one.

Let me chill.

10:55am. Let me finally start.

I am not really a rank 5. The pinnacle of programming skill would not possess such shallow accumulation. The pinnacle of skill even for just humans would have significant effect upon the world. I need to chase after it.

11am. Programming is not about conning my way into some company or the other. It is not about recognition or sponsors. It is about my own personal power.

The ML library was a mistake, but was I wrong to push forth with my own style? Never!

11:05am. Now I need to focus. I need to make the UI for the Leduc game.

11:25am. At some point I am going to have to make a serious effort at learning Python concurrency. Right now I really have no idea how do independent training on a separate thread. Compared to my F# skills with the Hopac library, I am just too lost in the Python ecosystem.

I just don't have the same kind of power here.

11:30am. Python concurrency and fixing the Spiral's editor support for the top level are my biggest challenge for the month March.

I need to get better at concurrency. 4/5 is not enough. I am barely 3/5 right now in Python.

I want to exercise this. This is the only thing worth doing in life. How could I ever want anything else?

...

Damn, Spiral v2 is awesome. February was my first real programming month with it and I am enjoying it. If only I could make time to redesign the top level so that those crashes when moving and deleting files stop happening it would be perfect.

I should do what I want.

11:35am. But for today, let me outline my goals.

1) I need to CPS the Leduc game a step further. Instead of the regular `game` nominal, I am going to do a version that accepts the return as a continuation. When it is the player's turn to act, I'll send a message to the UI with the continuation.

...

Hmmm...I know what I want, but I am not sure how to make the design fit.

11:40am. Well, it will take some work. The real problem is fitting it in with the previous design.

I'll leave this first goal for later.

2) I need to make the UI itself. This won't be too hard. I just need to expand the bindings and get more familiar with the Kivy library.

11:45am. I am thinking about this right now. Let me just give it a try.

I need to play around with Kivy widgets. What do I use for diplaying text? What about separators?

11:45am. https://kivy.org/doc/stable/guide/widgets.html

There is actually a separator used here in the catalogue example.

12pm. I am trying to make some concrete goals. It is good that I have a plan at the top level, but what are the concrete steps that I should do now?

12:05pm. I feel like forgetting the game and the language server issues and just grinding Kivy.

Since I will be using Kivy, I should go through the examples and translate some of them to Spiral. I should also fill out the bindings for the primitives.

Right now I only have a single layout and don't even understand how to do something as simple as a separator. I should take it easy and play with the library for a while. Get to know it while at the same time writing bindings and replicating the examples in Spiral.

This would be different than the furious outburst that I had imagined, but I should build my foundation in Python slowly.

Since I really want to, I should start planning how to redesign the top level of editor support. I want to avoid this as it would be very time consuming, but having the language be 100% stable shouldn't be an auxiliary goal. I should not blame myself for taking a detour. Even if it took me a few weeks so what.

When I am fully satisfied with language I will be able to rest more easily. I do not want to move a few files and think: 'Shit, now I have to restart the plugin.'

12:20pm. Do I want to play with Kivy or work on editor support?

I cannot do both.

12:25pm. Sigh. The later. I have to do the later. Better to do it now and enjoy the benefits of it throughout my work, than leave it for some indefinite future. Right now, it feels like Spiral's front porch has a pileup of garbage instead of the welcome mat.

February was really fun. I had enjoyed programming in Spiral itself after working towards this goal through the entirety of 2020 and the January of 2021. I did some non-trivial things and fixed what were the likely the last remaining bugs in the compiler.

I want the top level to be better. I am afraid that this issue might consume the entirety of March for me. But I should just face my fear.

At this point in my development I am frequently moving files around in the editor. I am not making single module tests anymore. I want the editor support to be roboust. The reason it is not is because my design is hacky and poor.

12:35pm. Let me have breakfast here, then I will go through the language from the top and identify all the pain points."