"```fs type PackageFilesTree =     | File of module_id: int * path : string * name: string option     | Directory of dir_id: int * path : string * name: string * PackageFilesTree list
type PackageFiles = { tree : PackageFilesTree list; num_dirs : int; num_modules : int }
```

Did a few edits in preparation for tomorrow. Having the relevant info be included up front will be of help down the road.

Knowing the numer of directories and modules up front will make it a lot easier to make the input array for updating the project files. I am going to give this all the care it deserves. This is why the slow and steady approach of building up the motivation and letting the inspiration come to you works.

I will savor this experience of March. When it is over, I am going to get back to work on ML. I won't be impatient to get back to ML or work on UIs.

After I deal with editor support and Spiral is back into operational state, I am going to take some time to collect all the contact info of various AI company contacts and fire them off to a dozen. If nothing comes of that then I will wash my hands of that affair. It is not core to my plan anyway."