"11:50am. Done. I guess I'll stop the morning session here. It was unusually light. I feel like I could have pushed harder rather than dwell in my own thoughts so much, but it is fine.
What should my next goal be? I think I should make a neural and an uniform player and test out both the vanilla and the CPS version on them. Command line of course.

I should that little bit out of the way and not dwell on it too much. After that I'll be able to move on to UIs.

11:55am. I am really pleased how things turned out. Especially those combinator functions `choice` and `iter`.

I just have to keep going like this, and the challenge will succumb to my might.

1:30pm. Let me do the chores here.

1:35pm. Whops, I just realized I forgot to publish the patch.

2pm. Done with chores. Let me catch my breath a little and I will start.

2:25pm. Let me start."