"9:55am. Let me chill a bit and I will start. I think I understand what my problem is now and where the dissonance is.
10:50am. Right now I am watching the Processing tuts. Let me elaborate what the problem is. The source of the dissonance is the difference between imagining an image and actually drawing it.

In the past week, I've thought about the data flow in depth. This is my expertise. I am good at managing and simplifying complex currents of data.

I've made the Lithe library and am pretending things are easy now, but that is not true. Things are only starting at this point.

Once the library is there, and it is not - these recent examples which refer to one another's properties are blowing a hole in my conception, once the library is there, I need to work on the UI in isolation.

Yesterday I asked whether I should write a game intepreter for the states.

Why am I even asking myself that? Really that is the easy part.

The hard part is working on the UI.

I should make and test all the components in isolation first. That is what I should be doing. It will be easy to ferry and transform the data once that is done. So what if I have to write a bit extra code to do it.

11am. What bothers me the most right now is how nasty the interop is.

In Spiral I do not get the benefit of Intellisense, and I had not anticipated some of the needs of the library.

I need to make more thorough bindings. Instead of just setters, I'll need both setters and getters. Let me resume the processing videos.

https://processing.org/tutorials/

Let me go through the text tutorials. I just want to do this. I want inspiration for my UI work. Maybe I'll download Processing for Python later.

https://processing.org/tutorials/overview/

> At the intersection of these points is a tradeoff between speed and simplicity of use. If we didn't care about speed, it might make sense to use Python, Ruby, or many other scripting languages. This is especially true for the education side. If we didn't care about making a transition to more advanced languages, we'd probably avoid a C++ or Java-style syntax. But Java makes a nice starting point for a sketching language because it's far more forgiving than C/C++ and also allows users to export sketches for distribution via the web.

So that is how it is.

11:40am. https://processing.org/tutorials/interactivity/

I am just skimming these right now to get a sense of it. I'll do some programming Spiral after the break. I am still thinking about things on the side.

11:50am. https://processing.org/tutorials/typography/

This is of interest to me. So far, Processing seems like half of UI library. I really am interested in what it offers for advanced use cases.

12:05pm. https://processing.org/tutorials/text/

Processing is based on an old version of Java with manifest typing.

12:15pm. https://processing.org/tutorials/pixels/

I did not think the tutorials would cover such low level details.

12:45pm. Ok, I get what Processing is about. I'll have to investigate the drawing capabilities of the Kivy library later.

Let me stop here for the time being.

I am in the right mindset to start drawing things. I'll play with the Kivy canvas and animations next.

I need to up my affinity with this. I need to make proper UIs, but for that I need to stop thinking about the backend stuff. So far this has been my focus, but this is not what I should be thinking about now that my focus are UIs themselves."