"3:15pm. Done with chores. Let me catch my breath and I will start.
Periods like these are the closest I'll get to having a vacation. I am going to pick up my pace when I feel like I've gotten all the basics out of the way. There is no need to agonize here.

3:20pm. Let me start. First comes the logo. Since it is simple it might be worth to put it as an `.svg`. But...I'll have to check out whether markdown supports it. Let me first find out how images can be embedded into markdown.

https://stackoverflow.com/questions/14494747/how-to-add-images-to-readme-md-on-github

Thank you SO. Oh right. Does it support `.svg`?

https://stackoverflow.com/questions/13808020/include-an-svg-hosted-on-github-in-markdown

Oh, it says that it works. Ok nice. I'll give it a try.

https://processing.org/reference/PShape_setFill_.html

Ah, maybe I can use this to set the fill and the stroke of already existing shapes. That might work.

> After using createShape(), stroke and fill color can be set by calling methods like setFill() and setStroke(), as seen in the examples above. The complete list of methods and fields for the PShape class are in the Processing Javadoc.

Yeah, this is common work, but I can't do it quickly. I got way too pessimistic about the icons. I guess it is not a wreck just yet. I'll have to go back to the example and make it work.

square = createShape(RECT, 0, 0, 50, 50);

It is possible to create rectangle shapes like this. This will give me a way to export those tables.

How do I save the shapes?

https://forum.processing.org/two/discussion/8007/how-to-save-pshapes

This is a really strange oversight. Processing has function to load shapes, but not to save them in the standard library.

The library says it is not 2.0 compatible.

https://processing.org/reference/libraries/svg/index.html

https://processing.org/reference/libraries/

Ah, maybe these are a part of the standard library now.

```java
import processing.svg.*;

void setup() {
  size(500,500,SVG,"spiral_logo.svg");
  //size(500,500,P2D);
  background(255);

  int num_segments = 30;
  int num_rotations = 3;
  float sc = 100;
  translate(height/2,width/2);
  scale(sc);
  noFill();
  PShape p1 = createShape();
  p1.beginShape();
  p1.strokeWeight(8/sc);
  PVector a = new PVector(-1.0 / 3.0, -1.0);
  PVector right = PVector.mult(PVector.fromAngle(0),1.0/3.0);
  PVector bCp1 = PVector.add(a,right);
  PVector bCp2 = PVector.add(bCp1,PVector.mult(PVector.fromAngle(radians(15)),2.5/3.0));
  PVector b = new PVector(1.0, 0.0);
  PVector cCp1 = PVector.add(PVector.mult(PVector.sub(b,bCp2),4.0/3.0),b);
  PVector cCp2 = PVector.add(cCp1,PVector.mult(right,1.0/2.0));
  PVector c = new PVector(5.0 / 3.0, 3.0 / 5.0);
  p1.vertex(a.x,a.y);
  p1.bezierVertex(bCp1.x,bCp1.y,bCp2.x,bCp2.y,b.x,b.y);
  p1.bezierVertex(cCp1.x,cCp1.y,cCp2.x,cCp2.y,c.x,c.y);
  p1.endShape();
  PShape p2 = createShape();
  p2.beginShape();
  p2.strokeWeight(4/sc);
  for (int i=0; i <= num_segments * num_rotations; i++) {
    float r = float (num_segments - i) / float (num_segments);
    float t = float (num_segments * num_rotations - i) / float (num_segments * num_rotations);
    p2.vertex(t * cos(TAU * r), t * sin(PI + TAU * r));
  }
  p2.endShape();
  shape(p1);
  shape(p2);
  //PShape all = createShape(GROUP);
  //all.addChild(p1);
  //all.addChild(p2);
  //shape(all);
}

void draw() {

}
```

Huh, it is telling me that the textMode(Shape) is not supported by the renderer. What is going on?

```java
import processing.svg.*;

void setup() {
  size(275,215,SVG,"spiral_logo.svg");
  //size(275,215,P2D);
  background(255);
}

void draw() {
  int num_segments = 30;
  int num_rotations = 3;
  float sc = 100;

  scale(sc);
  translate(1,1.1);
  noFill();

  beginShape();
  strokeWeight(8/sc);
  PVector a = new PVector(-1.0 / 3.0, -1.0);
  PVector right = PVector.mult(PVector.fromAngle(0),1.0/3.0);
  PVector bCp1 = PVector.add(a,right);
  PVector bCp2 = PVector.add(bCp1,PVector.mult(PVector.fromAngle(radians(15)),2.5/3.0));
  PVector b = new PVector(1.0, 0.0);
  PVector cCp1 = PVector.add(PVector.mult(PVector.sub(b,bCp2),4.0/3.0),b);
  PVector cCp2 = PVector.add(cCp1,PVector.mult(right,1.0/2.0));
  PVector c = new PVector(5.0 / 3.0, 3.0 / 5.0);
  vertex(a.x,a.y);
  bezierVertex(bCp1.x,bCp1.y,bCp2.x,bCp2.y,b.x,b.y);
  bezierVertex(cCp1.x,cCp1.y,cCp2.x,cCp2.y,c.x,c.y);
  endShape();

  beginShape();
  strokeWeight(4/sc);
  for (int i=0; i <= num_segments * num_rotations; i++) {
    float r = float (num_segments - i) / float (num_segments);
    float t = float (num_segments * num_rotations - i) / float (num_segments * num_rotations);
    vertex(t * cos(TAU * r), t * sin(PI + TAU * r));
  }
  endShape();

  println("Finished.");
  exit();
}
```

Given that the SVG library cannot deal with shapes properly, I think all of this is good incentive for me to not actually use shapes.

```java
//import processing.svg.*;

void setup() {
  size(400, 400, SVG, "filename.svg");
}

void draw() {
  // Draw something good here
  line(0, 0, width/2, height);

  // Exit the program
  println("Finished.");
  exit();
}
```

Let me just play around with this for a bit. If shapes cannot be used directly by this rendered, then what will I do when the time comes to load them?

`textMode(SHAPE) is not supported by this renderer.`

This just floors me.

https://forum.processing.org/two/discussion/26900/rendering-text-in-svg-files

> The SVG is being saved, however the text() is being rendered as a text tag instead of a Path, which is the behavior I would like and that happens to be the case with the PDF Renderer.

```java
import processing.svg.*;

void setup() {
  size(400, 400, SVG, "filename.svg");
  //size(400, 400);
}

void draw() {
  PShape s = createShape();
  s.beginShape();
  s.fill(255,0,0);
  s.vertex(0,0);
  s.vertex(50,0);
  s.vertex(50,50);
  s.endShape(CLOSE);
  shape(s);

  // Exit the program
  println("Finished.");
  exit();
}
```

I get the error message, but this does in fact work. The correct file does get outputted. How confusing.

Let me report the error for this.

4:25pm. Opened an issue. Let me move on. Time to put the Spiral logo into the project folder.

I've put it in. ...It works in preview. Let me try pushing it and viewing it online."